Intermediate changes for Qutron backend
