+++ docker/pythonpath_dev/superset_config.py
@@ -108,7 +108,8 @@ WEBDRIVER_BASEURL = "http://superset:8088/"
 WEBDRIVER_BASEURL_USER_FRIENDLY = WEBDRIVER_BASEURL
 WTF_CSRF_ENABLED=False
 SQLLAB_CTAS_NO_LIMIT = True
-
+SUPERSET_WEBSERVER_PROTOCOL = "https"
+ENABLE_PROXY_FIX = True
 #
 # Optionally import superset_config_docker.py (which will have been included on
 # the PYTHONPATH) in order to allow for local settings to be overridden
