+++ superset/quotron/DataTypes.py
@@ -1,0 +1,10 @@
+
+class Autocomplete:
+    def __init__(self, question, email, time):
+        self.question = question
+        self.email = email
+        self.time = time
+
+
+
+
